+++ blueprint/implementations/worker/bullmq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @forklaunch/implementation-worker-bullmq
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes