+++ blueprint/implementations/worker/kafka/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @forklaunch/implementation-worker-kafka
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes